requirements.txt
gdown>=5.2.0
deepgram-sdk>=3.7.0
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.60.0
google-auth>=2.29.0
ffmpeg-python>=0.2.0
requests>=2.31.0
click>=8.1.0
rich>=13.7.0
pytest>=8.0.0
pytest-mock>=3.12.0
